fix: add explicit cast in as_num() to resolve MSVC C4244 narrowing conversion error
yyjson_get_sint/yyjson_mut_get_sint returns int64_t, which implicitly
narrows to double when constructing std::optional<double>. MSVC treats
C4244 as error under /W4 /WX, breaking compilation.

Add static_cast<double>() to both const_value_ref and mutable
const_value_ref overloads.

+++ include/cpp_yyjson.hpp
@@ -1950,7 +1950,7 @@ namespace yyjson
                 {
                     if (is_real()) return yyjson_mut_get_real(base::val_);
                     if (is_int()) [[likely]]
-                        return yyjson_mut_get_sint(base::val_);
+                        return static_cast<double>(yyjson_mut_get_sint(base::val_));
                     return std::nullopt;
                 }
                 [[nodiscard]] std::optional<std::string_view> as_string() const noexcept
@@ -3503,7 +3503,7 @@ namespace yyjson
             {
                 if (is_real()) return yyjson_get_real(val_);
                 if (is_int()) [[likely]]
-                    return yyjson_get_sint(val_);
+                    return static_cast<double>(yyjson_get_sint(val_));
                 return std::nullopt;
             }
             [[nodiscard]] std::optional<std::string_view> as_string() const noexcept
